Add files via upload
The Small Office Home Office (SOHO) project aims to design and implement a network  using Cisco Packet Tracer for a small office setup. The network is intended to support the  networking needs of a small office or home office environment, with emphasis on  scalability, security, and cost-effectiveness.

The network design includes the use of Cisco Packet Tracer, a powerful simulation tool that  allows for the creation of virtual networks and the testing of network configurations in a  safe and controlled environment. The network consists of multiple devices, including  routers, switches, PCs, and servers, all configured using Cisco Packet Tracer.

The project focuses on meeting the networking requirements of a typical SOHO  environment, including the establishment of local area network (LAN) connectivity,  internet access, data segregation, security measures, and network management. VLANs are  configured to separate different departments or users, and appropriate routing protocols and  security features are implemented to ensure data integrity and confidentiality. The network  is also designed to be scalable, allowing for future expansion or changes in the network  requirements.

Security measures, such as firewalls, access controls, and intrusion detection and  prevention systems, are implemented to protect the network from potential threats. Cost-  effective solutions, such as choosing appropriate Cisco networking equipment and  configurations, are considered to meet the budget constraints of a small office or home  office setup.

The project is implemented and tested using Cisco Packet Tracer, with network  configurations and simulations performed to validate the design and functionality of the  SOHO network. The network is also managed and monitored using appropriate network  management tools within Cisco Packet Tracer